+++ src/main/java/com/troblecodings/invisiblelights/blocks/BlockCustomLight.java
@@ -2,8 +2,7 @@ package com.troblecodings.invisiblelights.blocks;
 
 public class BlockCustomLight extends BlockInvisibleLight {
 
-    public BlockCustomLight(final int light) {
-        super(light);
-    }
-
+	public BlockCustomLight(final int light) {
+		super(light);
+	}
 }
