+++ src/main/java/com/troblecodings/invisiblelights/blocks/BlockCustomState.java
@@ -3,70 +3,55 @@ package com.troblecodings.invisiblelights.blocks;
 import java.util.Random;
 
 import net.minecraft.block.Block;
-import net.minecraft.block.properties.PropertyBool;
-import net.minecraft.block.state.BlockStateContainer;
-import net.minecraft.block.state.IBlockState;
+import net.minecraft.block.BlockState;
+import net.minecraft.state.BooleanProperty;
+import net.minecraft.state.StateContainer;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
+import net.minecraft.world.server.ServerWorld;
 
 public class BlockCustomState extends BlockCustomLight {
 
-    public static final PropertyBool POWERED = PropertyBool.create("powered");
+	public static final BooleanProperty POWERED = BooleanProperty.create("powered");
 
-    public BlockCustomState(final int light) {
-        super(light);
-    }
+	public BlockCustomState(final int light) {
+		super(light);
+		setDefaultState(this.stateContainer.getBaseState().with(POWERED, false));
+	}
 
-    @Override
-    public IBlockState getStateFromMeta(final int meta) {
-        return this.getDefaultState().withProperty(POWERED, meta == 0 ? false : true);
-    }
+	@Override
+	protected void fillStateContainer(final StateContainer.Builder<Block, BlockState> builder) {
+		builder.add(POWERED);
+	}
 
-    @Override
-    public int getMetaFromState(final IBlockState state) {
-        return state.getValue(POWERED) ? 1 : 0;
-    }
+	private void updatePowerState(final BlockState state, final World worldIn, final BlockPos pos) {
+		if (worldIn.isRemote)
+			return;
+		final boolean lastPowered = state.get(POWERED);
+		if (worldIn.isBlockPowered(pos) && !lastPowered) {
+			worldIn.setBlockState(pos, state.with(POWERED, true), 3);
+		} else if (!worldIn.isBlockPowered(pos) && lastPowered) {
+			worldIn.getPendingBlockTicks().scheduleTick(pos, this, 4);
+		}
+	}
 
-    @Override
-    protected BlockStateContainer createBlockState() {
-        return new BlockStateContainer(this, POWERED);
-    }
+	@Override
+	public void onBlockAdded(final BlockState state, final World worldIn, final BlockPos pos,
+			final BlockState oldState, final boolean isMoving) {
+		super.onBlockAdded(state, worldIn, pos, oldState, isMoving);
+		updatePowerState(state, worldIn, pos);
+	}
 
-    @Override
-    public void onBlockAdded(final World worldIn, final BlockPos pos, final IBlockState state) {
-        super.onBlockAdded(worldIn, pos, state);
-        if (worldIn.isRemote)
-            return;
-        final boolean lastPowered = state.getValue(POWERED);
-        if (worldIn.isBlockPowered(pos) && !lastPowered) {
-            worldIn.setBlockState(pos, state.withProperty(POWERED, true), 3);
-        } else if (!worldIn.isBlockPowered(pos) && lastPowered) {
-            worldIn.scheduleUpdate(pos, this, 4);
-        }
-    }
+	@Override
+	public void neighborChanged(final BlockState state, final World worldIn, final BlockPos pos,
+			final Block blockIn, final BlockPos fromPos, final boolean isMoving) {
+		updatePowerState(state, worldIn, pos);
+	}
 
-    @Override
-    public void updateTick(final World worldIn, final BlockPos pos, final IBlockState state,
-            final Random rand) {
-        super.updateTick(worldIn, pos, state, rand);
-        if (worldIn.isRemote)
-            return;
-        final boolean lastPowered = state.getValue(POWERED);
-        if (!worldIn.isBlockPowered(pos) && lastPowered) {
-            worldIn.setBlockState(pos, state.withProperty(POWERED, false), 3);
-        }
-    }
-
-    @Override
-    public void neighborChanged(final IBlockState state, final World worldIn, final BlockPos pos,
-            final Block blockIn, final BlockPos fromPos) {
-        if (worldIn.isRemote)
-            return;
-        final boolean lastPowered = state.getValue(POWERED);
-        if (worldIn.isBlockPowered(pos) && !lastPowered) {
-            worldIn.setBlockState(pos, state.withProperty(POWERED, true), 3);
-        } else if (!worldIn.isBlockPowered(pos) && lastPowered) {
-            worldIn.scheduleUpdate(pos, this, 4);
-        }
-    }
+	@Override
+	public void tick(final BlockState state, final ServerWorld worldIn, final BlockPos pos, final Random rand) {
+		if (!worldIn.isBlockPowered(pos) && state.get(POWERED)) {
+			worldIn.setBlockState(pos, state.with(POWERED, false), 3);
+		}
+	}
 }
