+++ src/main/java/com/troblecodings/invisiblelights/blocks/BlockLightBlocker.java
@@ -1,33 +1,17 @@
 package com.troblecodings.invisiblelights.blocks;
 
-import com.troblecodings.invisiblelights.init.ILInit;
-
-import net.minecraft.block.state.IBlockState;
-import net.minecraft.util.EnumFacing;
+import net.minecraft.block.BlockState;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.IBlockAccess;
+import net.minecraft.world.IBlockReader;
 
 public class BlockLightBlocker extends BlockInvisibleLight {
 
-    public BlockLightBlocker() {
-        super(0);
-        this.setCreativeTab(ILInit.LIGHT_TAB);
-        this.disableStats();
-    }
+	public BlockLightBlocker() {
+		super(0);
+	}
 
-    @Override
-    public boolean isFullBlock(final IBlockState state) {
-        return false;
-    }
-
-    @Override
-    public boolean isOpaqueCube(final IBlockState state) {
-        return true;
-    }
-
-    @Override
-    public boolean doesSideBlockRendering(final IBlockState state, final IBlockAccess world,
-            final BlockPos pos, final EnumFacing face) {
-        return false;
-    }
+	@Override
+	public int getOpacity(final BlockState state, final IBlockReader worldIn, final BlockPos pos) {
+		return 15;
+	}
 }
