+++ src/main/java/com/troblecodings/invisiblelights/proxy/ClientProxy.java
@@ -2,203 +2,167 @@ package com.troblecodings.invisiblelights.proxy;
 
 import java.util.ArrayList;
 
+import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.vertex.IVertexBuilder;
 import com.troblecodings.invisiblelights.blocks.BlockCustomLight;
 import com.troblecodings.invisiblelights.blocks.BlockInvisibleLight;
 import com.troblecodings.invisiblelights.blocks.BlockLightBlocker;
 import com.troblecodings.invisiblelights.init.ILInit;
-import com.troblecodings.invisiblelights.init.ILModel;
 
 import net.minecraft.block.Block;
-import net.minecraft.block.state.IBlockState;
+import net.minecraft.block.BlockState;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.entity.EntityPlayerSP;
-import net.minecraft.client.renderer.GlStateManager;
-import net.minecraft.client.renderer.RenderGlobal;
-import net.minecraft.client.renderer.block.model.ModelResourceLocation;
+import net.minecraft.client.entity.player.ClientPlayerEntity;
+import net.minecraft.client.renderer.IRenderTypeBuffer;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.RenderTypeLookup;
+import net.minecraft.client.renderer.WorldRenderer;
 import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.Item;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.util.math.AxisAlignedBB;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.BlockRayTraceResult;
 import net.minecraft.util.math.RayTraceResult;
+import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.World;
-import net.minecraftforge.client.event.DrawBlockHighlightEvent;
-import net.minecraftforge.client.event.ModelRegistryEvent;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+import net.minecraftforge.client.event.DrawHighlightEvent;
 import net.minecraftforge.client.event.RenderWorldLastEvent;
-import net.minecraftforge.client.model.ModelLoader;
-import net.minecraftforge.client.model.ModelLoaderRegistry;
-import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.world.BlockEvent.BreakEvent;
 import net.minecraftforge.event.world.BlockEvent.EntityPlaceEvent;
-import net.minecraftforge.fml.common.event.FMLInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 
-public class ClientProxy extends CommonProxy {
+@OnlyIn(Dist.CLIENT)
+public final class ClientProxy {
 
-    @Override
-    public void preinit(final FMLPreInitializationEvent event) {
-        super.preinit(event);
-        MinecraftForge.EVENT_BUS.register(ClientProxy.class);
-        ModelLoaderRegistry.registerLoader(new ILModel());
-    }
+	private ClientProxy() {}
 
-    @SubscribeEvent
-    public static void modelEvents(final ModelRegistryEvent event) {
-        ILInit.ITEMS_TO_REGISTER.forEach(item -> ModelLoader.setCustomModelResourceLocation(item, 0,
-                new ModelResourceLocation(item.getRegistryName(), "inventory")));
-        ILInit.BLOCKS_TO_REGISTER.forEach(block -> {
-            final Item item = Item.getItemFromBlock(block);
-            ModelLoader.setCustomModelResourceLocation(item, 0,
-                    new ModelResourceLocation(item.getRegistryName(), "inventory"));
-        });
+	public static void onClientSetup(final FMLClientSetupEvent event) {
+		RenderTypeLookup.setRenderLayer(ILInit.GHOST_GLOWSTONE, RenderType.getCutoutMipped());
+	}
 
-    }
+	private static final int RADIUS = 50;
+	private static final int UPDATE_SPHERE = 50;
+	private static final int RADIUSPLAYER = RADIUS * RADIUS + 10;
 
-    private static final int RADIUS = 50;
-    private static final int UPDATE_SPHERE = 50;
-    private static final int RADIUSPLAYER = RADIUS * RADIUS + 10;
-    private static double d1;
-    private static double d2;
-    private static double d3;
+	private static final float[] COLOR_NORMAL = new float[] { 0, 1, 0, 1 };
+	private static final float[] COLOR_BLOCKER = new float[] { 1, 0, 0, 1 };
+	private static final float[] COLOR_CUSTOM = new float[] { 1, 0.5f, 0, 1 };
 
-    private static final float[] COLOR_NORMAL = new float[] {
-            0, 1, 0, 1
-    };
-    private static final float[] COLOR_BLOCKER = new float[] {
-            1, 0, 0, 1
-    };
-    private static final float[] COLOR_CUSTOM = new float[] {
-            1, 0.5f, 0, 1
-    };
+	private static final ArrayList<BlockPos> playerPlacedBlocks = new ArrayList<>();
+	private static boolean dirty = true;
+	private static BlockPos lastPosition = BlockPos.ZERO;
 
-    public static void render(final BlockPos pos1, final float[] color) {
-        RenderGlobal.drawSelectionBoundingBox(Block.FULL_BLOCK_AABB.offset((pos1.getX()) - d1,
-                (pos1.getY()) - d2, (pos1.getZ()) - d3), color[0], color[1], color[2], color[3]);
-    }
+	public static void refill(final BlockPos pos, final World world) {
+		lastPosition = pos;
+		dirty = false;
+		new Thread(() -> {
+			for (int x = -RADIUS; x <= RADIUS; x++) {
+				for (int y = -RADIUS; y <= RADIUS; y++) {
+					for (int z = -RADIUS; z <= RADIUS; z++) {
+						final BlockPos nPos = pos.add(x, y, z);
+						final Block pBlock = world.getBlockState(nPos).getBlock();
+						if (pBlock instanceof BlockInvisibleLight) {
+							synchronized (playerPlacedBlocks) {
+								playerPlacedBlocks.add(nPos);
+							}
+						}
+					}
+				}
+			}
+		}).start();
+	}
 
-    private static ArrayList<BlockPos> playerPlacedBlocks = new ArrayList<>();
-    private static boolean dirty = true;
-    private static BlockPos lastPosition = BlockPos.ORIGIN;
+	@SubscribeEvent
+	public static void renderOverlayEvent(final DrawHighlightEvent.HighlightBlock event) {
+		final BlockRayTraceResult result = event.getTarget();
+		final PlayerEntity player = Minecraft.getInstance().player;
+		if (player == null)
+			return;
+		final World world = player.getEntityWorld();
+		if (world == null)
+			return;
+		final BlockState state = world.getBlockState(result.getPos());
+		if (state.getBlock() instanceof BlockInvisibleLight)
+			event.setCanceled(true);
+	}
 
-    public static void refill(final BlockPos pos, final World world) {
-        lastPosition = pos;
-        dirty = false;
-        new Thread(() -> {
-            for (int x = -RADIUS; x <= RADIUS; x++) {
-                for (int y = -RADIUS; y <= RADIUS; y++) {
-                    for (int z = -RADIUS; z <= RADIUS; z++) {
-                        final BlockPos nPos = pos.add(x, y, z);
-                        final Block pBlock = world.getBlockState(nPos).getBlock();
-                        if (pBlock instanceof BlockInvisibleLight) {
-                            synchronized (playerPlacedBlocks) {
-                                playerPlacedBlocks.add(nPos);
-                            }
-                        }
-                    }
-                }
-            }
-        }).start();
-    }
+	@SubscribeEvent
+	public static void blockPlaceEvent(final EntityPlaceEvent event) {
+		final Entity placerEntity = event.getEntity();
+		if (placerEntity == null)
+			return;
+		final ClientPlayerEntity player = Minecraft.getInstance().player;
+		if (player == null)
+			return;
+		final BlockPos playerPos = player.getPosition();
+		final double distance = placerEntity.getPosition().distanceSq(playerPos);
+		if (distance < RADIUSPLAYER) {
+			if (event.getPlacedBlock().getBlock() instanceof BlockInvisibleLight) {
+				playerPlacedBlocks.clear();
+				refill(playerPos, player.getEntityWorld());
+			}
+		}
+	}
 
-    @SideOnly(Side.CLIENT)
-    @SubscribeEvent
-    public static void renderOverlayEvent(final DrawBlockHighlightEvent render) {
-        final EntityPlayer player = render.getPlayer();
-        if (player == null)
-            return;
-        final World world = player.getEntityWorld();
-        if (world == null)
-            return;
-        final RayTraceResult result = render.getTarget();
-        if (result == null)
-            return;
-        final BlockPos pos = result.getBlockPos();
-        if (pos == null)
-            return;
-        final IBlockState state = world.getBlockState(pos);
-        if (state.getBlock() instanceof BlockInvisibleLight) {
-            render.setCanceled(true);
-        }
-    }
+	@SubscribeEvent
+	public static void blockBreakEvent(final BreakEvent event) {
+		synchronized (playerPlacedBlocks) {
+			playerPlacedBlocks.remove(event.getPos());
+		}
+	}
 
-    @SideOnly(Side.CLIENT)
-    @SubscribeEvent
-    public static void blockPlaceEvent(final EntityPlaceEvent event) {
-        final Entity placerEntity = event.getEntity();
-        if (placerEntity == null)
-            return;
-        final EntityPlayerSP player = Minecraft.getMinecraft().player;
-        if (player == null)
-            return;
-        final BlockPos playerPos = player.getPosition();
-        final double distance = placerEntity.getPosition().distanceSq(playerPos);
-        if (distance < RADIUSPLAYER) {
-            if (event.getPlacedBlock().getBlock() instanceof BlockInvisibleLight) {
-                playerPlacedBlocks.clear();
-                refill(playerPos, player.getEntityWorld());
-            }
-        }
-    }
+	@SubscribeEvent
+	public static void renderWorldLastEvent(final RenderWorldLastEvent event) {
+		final ClientPlayerEntity sp = Minecraft.getInstance().player;
+		if (sp == null)
+			return;
+		final Block block = Block.getBlockFromItem(sp.getHeldItemMainhand().getItem());
+		if (!(block instanceof BlockInvisibleLight)) {
+			if (!playerPlacedBlocks.isEmpty()) {
+				playerPlacedBlocks.clear();
+				dirty = true;
+			}
+			return;
+		}
 
-    @SideOnly(Side.CLIENT)
-    @SubscribeEvent
-    public static void blockPlaceEvent(final BreakEvent event) {
-        synchronized (playerPlacedBlocks) {
-            if (playerPlacedBlocks.contains(event.getPos())) {
-                playerPlacedBlocks.remove(event.getPos());
-            }
-        }
-    }
+		final BlockPos pos = sp.getPosition();
+		if (pos.distanceSq(lastPosition) > UPDATE_SPHERE) {
+			synchronized (playerPlacedBlocks) {
+				playerPlacedBlocks.clear();
+			}
+			dirty = true;
+		}
+		if (dirty)
+			refill(pos, sp.world);
+		if (playerPlacedBlocks.isEmpty())
+			return;
 
-    @SubscribeEvent
-    public static void modelEvents(final RenderWorldLastEvent event) {
-        final EntityPlayerSP sp = Minecraft.getMinecraft().player;
-        final Block block = Block.getBlockFromItem(sp.getHeldItemMainhand().getItem());
-        if (block instanceof BlockInvisibleLight) {
-            final BlockPos pos = sp.getPosition();
-            if (pos.distanceSq(lastPosition) > UPDATE_SPHERE) {
-                synchronized (playerPlacedBlocks) {
-                    playerPlacedBlocks.clear();
-                }
-                dirty = true;
-            }
-            if (dirty) {
-                refill(pos, sp.world);
-            }
-            if (playerPlacedBlocks.isEmpty())
-                return;
-            final double part = event.getPartialTicks();
-            d1 = sp.lastTickPosX + (sp.posX - sp.lastTickPosX) * part;
-            d2 = sp.lastTickPosY + (sp.posY - sp.lastTickPosY) * part;
-            d3 = sp.lastTickPosZ + (sp.posZ - sp.lastTickPosZ) * part;
+		final Vec3d view = Minecraft.getInstance().gameRenderer.getActiveRenderInfo().getProjectedView();
+		final MatrixStack matrixStack = event.getMatrixStack();
+		matrixStack.push();
+		matrixStack.translate(-view.x, -view.y, -view.z);
 
-            GlStateManager.disableTexture2D();
-            synchronized (playerPlacedBlocks) {
-                playerPlacedBlocks.forEach(posIn -> {
-                    final Block blockIn = sp.world.getBlockState(posIn).getBlock();
-                    final float[] color = blockIn instanceof BlockLightBlocker ? COLOR_BLOCKER
-                            : (blockIn instanceof BlockCustomLight ? COLOR_CUSTOM : COLOR_NORMAL);
-                    ClientProxy.render(posIn, color);
-                });
-            }
-            GlStateManager.enableTexture2D();
-        } else if (!playerPlacedBlocks.isEmpty()) {
-            playerPlacedBlocks.clear();
-            dirty = true;
-        }
-    }
+		final IRenderTypeBuffer.Impl buffers = Minecraft.getInstance().getRenderTypeBuffers().getBufferSource();
+		final IVertexBuilder builder = buffers.getBuffer(RenderType.getLines());
 
-    @Override
-    public void init(final FMLInitializationEvent event) {
-        super.init(event);
-    }
+		synchronized (playerPlacedBlocks) {
+			playerPlacedBlocks.forEach(posIn -> {
+				final Block blockIn = sp.world.getBlockState(posIn).getBlock();
+				final float[] color = blockIn instanceof BlockLightBlocker
+						? COLOR_BLOCKER
+						: (blockIn instanceof BlockCustomLight ? COLOR_CUSTOM : COLOR_NORMAL);
+				final AxisAlignedBB box = new AxisAlignedBB(
+						posIn.getX(), posIn.getY(), posIn.getZ(),
+						posIn.getX() + 1.0, posIn.getY() + 1.0, posIn.getZ() + 1.0);
+				WorldRenderer.drawBoundingBox(matrixStack, builder, box,
+						color[0], color[1], color[2], color[3]);
+			});
+		}
 
-    @Override
-    public void postinit(final FMLPostInitializationEvent event) {
-        super.postinit(event);
-    }
-
+		matrixStack.pop();
+		buffers.finish(RenderType.getLines());
+	}
 }
