--- src/main/java/eu/gir/basics/blocks/BlockCustomLight.java
@@ -1,8 +1,0 @@
-package eu.gir.basics.blocks;
-
-public class BlockCustomLight extends BlockInvisibleLight {
-
-	public BlockCustomLight(final int light) {
-		super(light);
-	}
-}
