--- src/main/java/eu/gir/basics/init/GIRInit.java
@@ -1,94 +1,0 @@
-package eu.gir.basics.init;
-
-import java.lang.reflect.Field;
-import java.lang.reflect.Modifier;
-import java.util.ArrayList;
-
-import eu.gir.basics.GIRMain;
-import eu.gir.basics.blocks.BlockGhostGlowstone;
-import eu.gir.basics.blocks.BlockInvisibleLight;
-import eu.gir.basics.blocks.BlockLightBlocker;
-import net.minecraft.block.Block;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.BlockItem;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemGroup;
-import net.minecraft.item.ItemStack;
-import net.minecraft.util.ResourceLocation;
-import net.minecraftforge.event.RegistryEvent;
-import net.minecraftforge.event.world.BlockEvent.BreakEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.registries.IForgeRegistry;
-
-public class GIRInit {
-
-	public static final ItemGroup LIGHT_TAB = new ItemGroup("invisiblelights") {
-		@Override
-		public ItemStack createIcon() {
-			return new ItemStack(INVISIBLE_LIGHTS_2);
-		}
-	};
-
-	public static final Block INVISIBLE_LIGHTS_2 = new BlockInvisibleLight(2);
-	public static final Block INVISIBLE_LIGHTS_3 = new BlockInvisibleLight(3);
-	public static final Block INVISIBLE_LIGHTS_4 = new BlockInvisibleLight(4);
-	public static final Block INVISIBLE_LIGHTS_5 = new BlockInvisibleLight(5);
-	public static final Block INVISIBLE_LIGHTS_6 = new BlockInvisibleLight(6);
-	public static final Block INVISIBLE_LIGHTS_7 = new BlockInvisibleLight(7);
-	public static final Block INVISIBLE_LIGHTS_8 = new BlockInvisibleLight(8);
-	public static final Block INVISIBLE_LIGHTS_9 = new BlockInvisibleLight(9);
-	public static final Block INVISIBLE_LIGHTS_10 = new BlockInvisibleLight(10);
-	public static final Block INVISIBLE_LIGHTS_11 = new BlockInvisibleLight(11);
-	public static final Block INVISIBLE_LIGHTS_12 = new BlockInvisibleLight(12);
-	public static final Block INVISIBLE_LIGHTS_13 = new BlockInvisibleLight(13);
-	public static final Block INVISIBLE_LIGHTS_14 = new BlockInvisibleLight(14);
-	public static final Block INVISIBLE_LIGHTS_15 = new BlockInvisibleLight(15);
-	public static final Block BLOCKER = new BlockLightBlocker();
-	public static final BlockGhostGlowstone GHOST_GLOWSTONE = new BlockGhostGlowstone();
-
-	public static final ArrayList<Block> blocksToRegister = new ArrayList<>();
-
-	public static void init() {
-		for (final Field field : GIRInit.class.getFields()) {
-			final int modifiers = field.getModifiers();
-			if (!Modifier.isStatic(modifiers) || !Modifier.isFinal(modifiers) || !Modifier.isPublic(modifiers))
-				continue;
-			try {
-				final Object obj = field.get(null);
-				if (!(obj instanceof Block))
-					continue;
-				final Block block = (Block) obj;
-				final String name = field.getName().toLowerCase().replace("_", "");
-				block.setRegistryName(new ResourceLocation(GIRMain.MODID, name));
-				blocksToRegister.add(block);
-			} catch (IllegalArgumentException | IllegalAccessException e) {
-				GIRMain.LOG.error("Could not access field {}", field.getName(), e);
-			}
-		}
-	}
-
-	public static void registerBlocks(final RegistryEvent.Register<Block> event) {
-		final IForgeRegistry<Block> registry = event.getRegistry();
-		blocksToRegister.forEach(registry::register);
-	}
-
-	public static void registerItems(final RegistryEvent.Register<Item> event) {
-		final IForgeRegistry<Item> registry = event.getRegistry();
-		blocksToRegister.forEach(block -> {
-			final BlockItem item = new BlockItem(block, new Item.Properties().group(LIGHT_TAB));
-			item.setRegistryName(block.getRegistryName());
-			registry.register(item);
-		});
-	}
-
-	@SubscribeEvent
-	public static void blockBreakEven(final BreakEvent event) {
-		if (!(event.getState().getBlock() instanceof BlockInvisibleLight))
-			return;
-		final PlayerEntity player = event.getPlayer();
-		final Item item = player.getHeldItemMainhand().getItem();
-		if (!(Block.getBlockFromItem(item) instanceof BlockInvisibleLight)) {
-			event.setCanceled(true);
-		}
-	}
-}
